Split the parts into its own files. This way makes it easier to see the different parts that were rather buried in big files. Started creating a Request class so that request building and execution can be done in it instead of DioMixin doing everything. DioMixin "has God class complex" needs to delegate its work to more focused components. Some methods are way too big and do way too much. Also, I finding this refactoring too easy. I have feeling there are many test cases missing. But maybe Im just being paranoid. Once reduced big methods and classes to smaller components I will have a deeper look at the tests.
