+++ dio/lib/adapter.dart
@@ -1,2 +1,2 @@
-export 'src/adapters/default_http_client_adapter.dart'
+export 'src/client_adapters/default_http_client_adapter.dart'
     show DefaultHttpClientAdapter;
