+++ dio/lib/adapter_browser.dart
@@ -1,2 +1,2 @@
-export 'src/adapters/browser_http_client_adapter.dart'
+export 'src/client_adapters/browser_http_client_adapter.dart'
     show BrowserHttpClientAdapter;
