+++ dio/lib/src/client_adapters/export.dart
@@ -1,0 +1,1 @@
+export 'http_client_adapter.dart';
