+++ dio/lib/src/instances/instance_stub.dart
@@ -1,3 +1,4 @@
 import '../dio.dart';
+import '../options/base_options.dart';
 
 Dio createDio([BaseOptions? options]) => throw UnsupportedError('');
