+++ dio/lib/src/options/base_options.dart
@@ -1,0 +1,93 @@
+import '../requests/request_config.dart';
+import '../requests/requests.dart';
+import '../responses/responses.dart';
+
+/// The common config for the Dio instance.
+/// `dio.options` is a instance of [BaseOptions]
+class BaseOptions extends RequestConfig {
+  BaseOptions({
+    String? method,
+    this.connectTimeout,
+    int? receiveTimeout,
+    int? sendTimeout,
+    this.baseUrl = '',
+    Map<String, dynamic>? queryParameters,
+    Map<String, dynamic>? extra,
+    Map<String, dynamic>? headers,
+    ResponseType? responseType = ResponseType.json,
+    String? contentType,
+    ValidateStatus? validateStatus,
+    bool? receiveDataWhenStatusError,
+    bool? followRedirects,
+    int? maxRedirects,
+    RequestEncoder? requestEncoder,
+    ResponseDecoder? responseDecoder,
+  })  : queryParameters = queryParameters ?? {},
+        super(
+          method: method,
+          receiveTimeout: receiveTimeout,
+          sendTimeout: sendTimeout,
+          extra: extra,
+          headers: headers,
+          responseType: responseType,
+          contentType: contentType,
+          validateStatus: validateStatus,
+          receiveDataWhenStatusError: receiveDataWhenStatusError,
+          followRedirects: followRedirects,
+          maxRedirects: maxRedirects,
+          requestEncoder: requestEncoder,
+          responseDecoder: responseDecoder,
+        );
+
+  /// Create a Option from current instance with merging attributes.
+  BaseOptions merge({
+    String? method,
+    String? baseUrl,
+    Map<String, dynamic>? queryParameters,
+    String? path,
+    int? connectTimeout,
+    int? receiveTimeout,
+    int? sendTimeout,
+    Map<String, dynamic>? extra,
+    Map<String, dynamic>? headers,
+    ResponseType? responseType,
+    String? contentType,
+    ValidateStatus? validateStatus,
+    bool? receiveDataWhenStatusError,
+    bool? followRedirects,
+    int? maxRedirects,
+    RequestEncoder? requestEncoder,
+    ResponseDecoder? responseDecoder,
+  }) {
+    return BaseOptions(
+      method: method ?? this.method,
+      baseUrl: baseUrl ?? this.baseUrl,
+      queryParameters: queryParameters ?? this.queryParameters,
+      connectTimeout: connectTimeout ?? this.connectTimeout,
+      receiveTimeout: receiveTimeout ?? this.receiveTimeout,
+      sendTimeout: sendTimeout ?? this.sendTimeout,
+      extra: extra ?? Map.from(this.extra),
+      headers: headers ?? Map.from(this.headers),
+      responseType: responseType ?? this.responseType,
+      contentType: contentType ?? this.contentType,
+      validateStatus: validateStatus ?? this.validateStatus,
+      receiveDataWhenStatusError:
+          receiveDataWhenStatusError ?? this.receiveDataWhenStatusError,
+      followRedirects: followRedirects ?? this.followRedirects,
+      maxRedirects: maxRedirects ?? this.maxRedirects,
+      requestEncoder: requestEncoder ?? this.requestEncoder,
+      responseDecoder: responseDecoder ?? this.responseDecoder,
+    );
+  }
+
+  /// Request base url, it can contain sub path, like: "https://www.google.com/api/".
+  late String baseUrl;
+
+  /// Common query parameters
+  Map<String, dynamic> queryParameters;
+
+  /// Timeout in milliseconds for opening url.
+  /// [Dio] will throw the [DioError] with [DioErrorType.CONNECT_TIMEOUT] type
+  ///  when time out.
+  int? connectTimeout;
+}
