+++ dio/lib/src/options/options.dart
@@ -1,0 +1,70 @@
+import '../requests/request_config.dart';
+import '../requests/requests.dart';
+import '../responses/responses.dart';
+
+/// Every request can pass an [Options] object which will be merged with [Dio.options]
+class Options extends RequestConfig {
+  Options({
+    String? method,
+    int? sendTimeout,
+    int? receiveTimeout,
+    Map<String, dynamic>? extra,
+    Map<String, dynamic>? headers,
+    ResponseType? responseType,
+    String? contentType,
+    ValidateStatus? validateStatus,
+    bool? receiveDataWhenStatusError,
+    bool? followRedirects,
+    int? maxRedirects,
+    RequestEncoder? requestEncoder,
+    ResponseDecoder? responseDecoder,
+  }) : super(
+          method: method,
+          sendTimeout: sendTimeout,
+          receiveTimeout: receiveTimeout,
+          extra: extra,
+          headers: headers,
+          responseType: responseType,
+          contentType: contentType,
+          validateStatus: validateStatus,
+          receiveDataWhenStatusError: receiveDataWhenStatusError,
+          followRedirects: followRedirects,
+          maxRedirects: maxRedirects,
+          requestEncoder: requestEncoder,
+          responseDecoder: responseDecoder,
+        );
+
+  /// Create a Option from current instance with merging attributes.
+  Options merge({
+    String? method,
+    int? sendTimeout,
+    int? receiveTimeout,
+    Map<String, dynamic>? extra,
+    Map<String, dynamic>? headers,
+    ResponseType? responseType,
+    String? contentType,
+    ValidateStatus? validateStatus,
+    bool? receiveDataWhenStatusError,
+    bool? followRedirects,
+    int? maxRedirects,
+    RequestEncoder? requestEncoder,
+    ResponseDecoder? responseDecoder,
+  }) {
+    return Options(
+      method: method ?? this.method,
+      sendTimeout: sendTimeout ?? this.sendTimeout,
+      receiveTimeout: receiveTimeout ?? this.receiveTimeout,
+      extra: extra ?? Map.from(this.extra),
+      headers: headers ?? Map.from(this.headers),
+      responseType: responseType ?? this.responseType,
+      contentType: contentType ?? this.contentType,
+      validateStatus: validateStatus ?? this.validateStatus,
+      receiveDataWhenStatusError:
+          receiveDataWhenStatusError ?? this.receiveDataWhenStatusError,
+      followRedirects: followRedirects ?? this.followRedirects,
+      maxRedirects: maxRedirects ?? this.maxRedirects,
+      requestEncoder: requestEncoder ?? this.requestEncoder,
+      responseDecoder: responseDecoder ?? this.responseDecoder,
+    );
+  }
+}
